backend/requirements.txt
# Python Backend Dependencies for Haulistry

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# GraphQL
strawberry-graphql[fastapi]==0.235.2

# Firebase Admin SDK
firebase-admin==6.5.0

# Neo4j Driver
neo4j==5.26.0

# Environment Variables
python-dotenv==1.0.1

# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# Password Hashing (backup/additional security)
passlib==1.7.4
bcrypt==4.2.0

# JWT Token Handling
# JWT Token Handling
python-jose[cryptography]==3.3.0
PyJWT==2.9.0

# HTTP Client
httpx==0.27.2

# Datetime utilities
python-dateutil==2.9.0

# Logging
python-json-logger==2.0.7

# Testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
